Future Detail done

+++ app/src/main/java/tsisyk/app/forecast/data/db/entity/Day.kt
@@ -1,32 +1,34 @@
 package tsisyk.app.forecast.data.db.entity
 
+import androidx.room.Embedded
+import com.google.gson.annotations.SerializedName
 
 data class Day(
-    @SerializedName("avgtemp_c")
+        @SerializedName("avgtemp_c")
     val avgtempC: Double,
-    @SerializedName("avgtemp_f")
+        @SerializedName("avgtemp_f")
     val avgtempF: Double,
-    @SerializedName("avgvis_km")
-    val avgvisKm: Int,
-    @SerializedName("avgvis_miles")
-    val avgvisMiles: Int,
-    @Embedded(prefix = "condition_")
+        @SerializedName("avgvis_km")
+    val avgvisKm: Double,
+        @SerializedName("avgvis_miles")
+    val avgvisMiles: Double,
+        @Embedded(prefix = "condition_")
     val condition: Condition,
-    @SerializedName("maxtemp_c")
+        @SerializedName("maxtemp_c")
     val maxtempC: Double,
-    @SerializedName("maxtemp_f")
+        @SerializedName("maxtemp_f")
     val maxtempF: Double,
-    @SerializedName("maxwind_kph")
+        @SerializedName("maxwind_kph")
     val maxwindKph: Double,
-    @SerializedName("maxwind_mph")
+        @SerializedName("maxwind_mph")
     val maxwindMph: Double,
-    @SerializedName("mintemp_c")
+        @SerializedName("mintemp_c")
     val mintempC: Double,
-    @SerializedName("mintemp_f")
+        @SerializedName("mintemp_f")
     val mintempF: Double,
-    @SerializedName("totalprecip_in")
-    val totalprecipIn: Int,
-    @SerializedName("totalprecip_mm")
-    val totalprecipMm: Int,
-    val uv: Double
+        @SerializedName("totalprecip_in")
+    val totalprecipIn: Double,
+        @SerializedName("totalprecip_mm")
+    val totalprecipMm: Double,
+        val uv: Double
 )
